+++ perf/src/utils/Perf.h
@@ -16,16 +16,16 @@ private:
 public:
 	Perf();
 	Perf(const Perf &p);
-	Perf(double ms);
+	Perf(float ms);
 	virtual ~Perf();
 
 	void start();
 	void stop();
 	void reset();
 
-	double getElapsedTime();                                                // ms
-	double getGflops(unsigned long flops);                                  // Gflops/s
-	double getMemoryBandwidth(unsigned long memops, unsigned short nBytes); // Go/s
+	float getElapsedTime();                                                // ms
+	float getGflops(float flops);                                          // Gflops/s
+	float getMemoryBandwidth(unsigned long memops, unsigned short nBytes); // Go/s
 
 	Perf operator+(const Perf& p);
 	Perf operator+=(const Perf& p);
